extract SiteIinfoDetail to SiteInfoDetailContents component
