+++ src/components/composite/shared-site-info/SharedSiteInfoDetail.tsx
@@ -2,6 +2,7 @@ import { View } from '../../Themed';
 import * as S from './SiteInfoDetail.styles';
 import { useCampSiteInfo } from '../../../api/site-info';
 import SiteInfoDetail from '../site-info/SiteInfoDetail';
+import SiteInfoDetailContents from '../site-info/SiteInfoDetailContents';
 
 interface SharedSiteInfoDetailProps {
   id: string;
@@ -18,7 +19,7 @@ const SharedSiteInfoDetail = ({ id }: SharedSiteInfoDetailProps) => {
           <S.Text>{siteInfo?.siteNumber}</S.Text>
         </View>
       </S.SiteInfoCardContainer>
-      <SiteInfoDetail id={id} />
+      {siteInfo && <SiteInfoDetailContents siteInfo={siteInfo} />}
     </S.Container>
   );
 };
